[tasks] add authorization

+++ task/internal/application/usecase/deletetask/deletetask.go
@@ -2,6 +2,7 @@ package deletetask
 
 import (
 	"context"
+	"task/internal/domain/models"
 	"task/internal/domain/models/task"
 )
 
@@ -10,6 +11,7 @@ type Usecase struct {
 }
 
 type TaskDeleter interface {
+	Task(ctx context.Context, taskId int64) (task.Task, error)
 	DeleteTask(ctx context.Context, taskId int64) error
 }
 
@@ -19,6 +21,20 @@ func New(deleter TaskDeleter) *Usecase {
 	}
 }
 
-func (u *Usecase) Delete(ctx context.Context, task *task.Task) error {
-	return u.deleter.DeleteTask(ctx, task.Id)
+func (u *Usecase) Delete(ctx context.Context, task *task.Task, userId int64) error {
+	t, err := u.deleter.Task(ctx, task.Id)
+	if err != nil {
+		return err
+	}
+
+	if !t.CanBeDeletedBy(userId) {
+		return models.ErrCannotDeleteRecord
+	}
+
+	err = u.deleter.DeleteTask(ctx, task.Id)
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
